rename EasifyConstants to EasifyDefines
