+++ Modules/CoreModules/EasifyCore/EasifyCore/Services/ServiceProvider.swift
@@ -20,7 +20,7 @@ public class ServiceProvider {
     // MARK: - Lifecycle
     public init() throws {
         do {
-            plistReaderService = try PlistReaderService(name: EasifyConstants.Files.spotifyPlistFileName,
+            plistReaderService = try PlistReaderService(name: EasifyDefines.Files.spotifyPlistFileName,
                                                         bundle: Bundle(for: type(of: self)))
             spotifyService = try SpotifyService(plistReaderService: plistReaderService)
         } catch {
